requirements.txt
pytest
tox
PyYAML
Jinja2
mypy